Update session when update profile/access constraint pending blog

+++ src/java/Controller/BlogDetailServlet.java
@@ -58,6 +58,9 @@ public class BlogDetailServlet extends HttpServlet {
         String url;
         String txtBlogID = request.getParameter("txtBlogID");
         int blogID;
+        HttpSession session = request.getSession();
+        AccountDTO account = (AccountDTO) session.getAttribute("USER");
+
         if (null != txtBlogID) {
             blogID = Integer.parseInt(txtBlogID);
         } else {
@@ -72,12 +75,20 @@ public class BlogDetailServlet extends HttpServlet {
                 AwardListDAO awardListDao = new AwardListDAO();
                 ArrayList<String> awawdNames = new ArrayList<>();
 
+                //Check pending and user constraint:
+                if (null==account||(!(account.getAccountID() == author.getAccountID()) && ("PENDING".equals(blog.getStatus())))) {
+                    System.out.println("Pending blog access denied");
+                    response.sendError(404);
+                    return;
+                }
+
+                //----------------------
                 //Remove expired rewards:
                 System.out.println("Blog ID: " + blogID);
                 if (awardListDao.deleteExpiredAwardFromBlogId(blogID)) {
                     System.out.println("Expired Award Cleared");
                 }
-
+                //---------------------
                 for (AwardDTO award : awardDao.getAllAward()) {
                     for (AwardListDTO awardList : awardListDao.getAwardListFromBlogId(blogID)) {
                         if (award.getAwardID() == awardList.getAwardID()) {
@@ -92,8 +103,6 @@ public class BlogDetailServlet extends HttpServlet {
                 request.setAttribute("AUTHOR", author);
 
                 // Check if already rated
-                HttpSession session = request.getSession();
-                AccountDTO account = (AccountDTO) session.getAttribute("USER");
                 if (null != account) {
                     BlogRatingDAO blogRatingDAO = new BlogRatingDAO();
                     BlogRatingDTO blogRatingDTO = blogRatingDAO.getBlogRatingFromBlogIDAndOwnerID(blogID, account.getAccountID());
